make MultiplexedTextWriter observable
