+++ src/Microsoft.DotNet.Interactive.Tests/Utility/MultiplexingTextWriterTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
 using System.IO;
 using FluentAssertions;
 using System.Linq;
@@ -10,6 +11,9 @@ using System.Threading;
 using System.Threading.Tasks;
 using FSharp.Compiler.Scripting;
 using Microsoft.CodeAnalysis.CSharp.Scripting;
+using Microsoft.CodeAnalysis.Text;
+using Microsoft.DotNet.Interactive.Commands;
+using Microsoft.DotNet.Interactive.Notebook;
 using Microsoft.DotNet.Interactive.Utility;
 using Microsoft.FSharp.Core;
 using Xunit;
@@ -78,23 +82,24 @@ namespace Microsoft.DotNet.Interactive.Tests.Utility
         [Fact]
         public void Disposal_of_outer_context_disposes_writer()
         {
-            var writer = new MultiplexingTextWriter(() => new TestWriter());
+            var innerWriter = new TestWriter();
+            var writer = new MultiplexingTextWriter(() => innerWriter);
             using var outer = writer.EnsureInitializedForCurrentAsyncContext();
             writer.Write("outer");
             var inner = writer.EnsureInitializedForCurrentAsyncContext();
             writer.Write(inner);
 
-            var testWriter = writer.Writers.OfType<TestWriter>().Single();
-
             outer.Dispose();
 
             writer.Writers.Count().Should().Be(0);
+            innerWriter.Disposed.Should().BeTrue();
         }
 
         [Fact]
         public void Disposal_of_inner_contexts_doe_not_dispose_writer()
         {
-            var writer = new MultiplexingTextWriter();
+            var innerWriter = new TestWriter();
+            var writer = new MultiplexingTextWriter(() => innerWriter);
             using var outer = writer.EnsureInitializedForCurrentAsyncContext();
             writer.Write("outer");
             var inner = writer.EnsureInitializedForCurrentAsyncContext();
@@ -102,11 +107,12 @@ namespace Microsoft.DotNet.Interactive.Tests.Utility
             
             inner.Dispose();
 
-            writer.Writers.Count().Should().Be(1);
+            writer.Writers.Count().Should().Be(1); 
+            innerWriter.Disposed.Should().BeFalse();
         }
 
         [Fact]
-        public async Task EXPERIMENT_async_context_is_available_in_csharp_scripting()
+        public async Task async_context_is_available_in_csharp_scripting()
         {
             var originalWriter = Console.Out;
 
@@ -117,7 +123,6 @@ namespace Microsoft.DotNet.Interactive.Tests.Utility
 
                 Console.SetOut(multiWriter);
 
-                // await kernel.SubmitCodeAsync("Console.WriteLine(\"hello!\");");
                 var result = await CSharpScript.RunAsync("System.Console.WriteLine(\"hello!\");");
 
                 result.Exception.Should().BeNull();
@@ -131,7 +136,7 @@ namespace Microsoft.DotNet.Interactive.Tests.Utility
         }
 
         [Fact]
-        public void EXPERIMENT_async_context_is_available_in_fsharp_scripting()
+        public void async_context_is_available_in_fsharp_scripting()
         {
             var originalWriter = Console.Out;
 
@@ -161,6 +166,107 @@ namespace Microsoft.DotNet.Interactive.Tests.Utility
                 Console.SetOut(originalWriter);
             }
         }
+
+        [Theory]
+        [MemberData(nameof(WriteOperations))]
+        public void Write_operations_on_MultiplexingStringWriter_are_observable_and_produce_one_event_per_top_level_write_invocation(
+            Action<TextWriter> write, 
+            string expectedValue)
+        {
+            var writer = new MultiplexingTextWriter();
+
+            using var _ = writer.EnsureInitializedForCurrentAsyncContext();
+
+            using var events = writer.GetObservable().ToSubscribedList();
+
+            write(writer);
+
+            events.Should()
+                  .ContainSingle()
+                  .Which
+                  .Should()
+                  .Be(expectedValue);
+        }
+
+        [Theory]
+        [MemberData(nameof(WriteOperations))]
+        public void Write_operations_on_ObservableStringWriter_are_observable_and_produce_one_event_per_top_level_write_invocation(
+            Action<TextWriter> write, 
+            string expectedValue)
+        {
+            var writer = new ObservableStringWriter();
+
+            using var events = writer.ToSubscribedList();
+
+            write(writer);
+
+            events.Should()
+                  .ContainSingle()
+                  .Which
+                  .Should()
+                  .Be(expectedValue);
+        }
+
+
+        public static IEnumerable<object[]> WriteOperations()
+        {
+            foreach (var (writeOperation, expectedValue) in actions())
+            {
+                yield return new object[] { writeOperation, expectedValue };
+            }
+
+            IEnumerable<( Action<TextWriter> writeOperation,  string expectedValue )> actions()
+            {
+                // yield return (writer => writer.Write("char[] buffer, int index, int count".ToArray()   , 0, 4), "char");
+                // yield return (writer => writer.Write("string value"), "string value");
+                // yield return (writer => writer.WriteAsync('c').Wait(), "c");
+                // yield return (writer => writer.WriteAsync(new[] { 'h', 'e', 'l', 'l', 'o' }, 0, 4).Wait(), "hell");
+                // yield return (writer => writer.WriteAsync("string value").Wait(), "string value");
+                // yield return (writer => writer.WriteLineAsync('c').Wait(), "c" + NewLine);
+                // yield return (writer => writer.WriteLineAsync("char[] buffer, int index, int count".ToArray(), 0, 4).Wait(), "char" + NewLine);
+                // yield return (writer => writer.WriteLineAsync("string value").Wait(), "string value" + NewLine);
+                // yield return (writer => writer.Write(true), true.ToString());
+                // yield return (writer => writer.Write("char[] buffer".ToArray()), "char[] buffer");
+                // yield return (writer => writer.Write(1d), 1d.ToString());
+                // yield return (writer => writer.Write(2.1), 2.1.ToString());
+                // yield return (writer => writer.Write(int.MaxValue), int.MaxValue.ToString());
+                // yield return (writer => writer.Write(long.MaxValue), long.MaxValue.ToString());
+                // yield return (writer => writer.Write(new object()), new object().ToString());
+                // yield return (writer => writer.Write(float.MaxValue), float.MaxValue.ToString());
+                // yield return (writer => writer.Write("string format, object {0}", 123), "string format, object 123");
+                // yield return (writer => writer.Write("string format, object {0}, object {1}", 456, 789), "string format, object 456, object 789");
+                // yield return (writer => writer.Write("string format, object {0}, object {1}, object {2}", 12, 34, 56), "string format, object 12, object 34, object 56");
+                // yield return (writer => writer.Write("string format, object {0}, object {1}, object {2}, object {3}", 123, false, new object(), 789), $"string format, object {123}, object {false}, object {new object()}, object {789}");
+                // yield return (writer => writer.Write(uint.MaxValue), uint.MaxValue.ToString());
+                // yield return (writer => writer.Write(ulong.MaxValue), ulong.MaxValue.ToString());
+                // yield return (writer => writer.WriteLine(), NewLine);
+                yield return (writer => writer.WriteLine(true), true + NewLine);
+                // yield return (writer => writer.WriteLine('c'), "c" + NewLine);
+                // yield return (writer => writer.WriteLine("char[] buffer".ToArray()), "char[] buffer" + NewLine);
+                // yield return (writer => writer.WriteLine("char[] buffer, int index, int count".ToArray(), 0, 4), "char" + NewLine);
+                // yield return (writer => writer.WriteLine(decimal.MaxValue), decimal.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine(double.MaxValue), double.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine(int.MaxValue), int.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine(long.MaxValue), long.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine(new object()), new object() + NewLine);
+                // yield return (writer => writer.WriteLine(float.MaxValue), float.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine("string value"), "string value" + NewLine);
+                //
+                // yield return (writer => writer.WriteLine("string format, object {0}", 123), "string format, object 123" + NewLine);
+                // yield return (writer => writer.WriteLine("string format, object {0}, object {1}", 456, 789), "string format, object 456, object 789" + NewLine);
+                // yield return (writer => writer.WriteLine("string format, object {0}, object {1}, object {2}", 12, 34, 56), "string format, object 12, object 34, object 56" + NewLine);
+                // yield return (writer => writer.WriteLine("string format, object {0}, object {1}, object {2}, object {3}", 123, false, new object(), 789), $"string format, object {123}, object {false}, object {new object()}, object {789}" + NewLine);
+                //
+                // yield return (writer => writer.WriteLine(uint.MaxValue), uint.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLine(ulong.MaxValue), ulong.MaxValue + NewLine);
+                // yield return (writer => writer.WriteLineAsync().Wait(), NewLine);
+                // yield return (writer => writer.Write("ReadOnlySpan<char> buffer".AsSpan()), "ReadOnlySpan<char> buffer");
+                // yield return (writer => writer.WriteAsync("ReadOnlyMemory<char> buffer, CancellationToken cancellationToken".AsMemory(), CancellationToken.None ).Wait(), "ReadOnlyMemory<char> buffer, CancellationToken cancellationToken");
+                // yield return (writer => writer.WriteLine("ReadOnlySpan<char> buffer".AsSpan()), "ReadOnlySpan<char> buffer" + NewLine);
+                // yield return (writer => writer.WriteLineAsync("ReadOnlyMemory<char> buffer, CancellationToken cancellationToken".AsMemory(), CancellationToken.None ).Wait(), "ReadOnlyMemory<char> buffer, CancellationToken cancellationToken" + NewLine);
+            }
+        }
+
 
         private class TestWriter : TextWriter
         {
